+++ internal/provider/data_source_iosxe_interface_vlan.go
@@ -88,6 +88,10 @@ func (d *InterfaceVLANDataSource) Schema(ctx context.Context, req datasource.Sch
 				MarkdownDescription: "Enable proxy ARP",
 				Computed:            true,
 			},
+			"ip_local_proxy_arp": schema.BoolAttribute{
+				MarkdownDescription: "Enable local-proxy ARP",
+				Computed:            true,
+			},
 			"ip_redirects": schema.BoolAttribute{
 				MarkdownDescription: "Enable sending ICMP Redirect messages",
 				Computed:            true,