+++ internal/provider/data_source_iosxe_interface_vlan_test.go
@@ -40,6 +40,7 @@ func TestAccDataSourceIosxeInterfaceVLAN(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "description", "My Interface Description"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "shutdown", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_proxy_arp", "false"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_local_proxy_arp", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_redirects", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_unreachables", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "vrf_forwarding", "VRF1"))
@@ -108,6 +109,7 @@ func testAccDataSourceIosxeInterfaceVLANConfig() string {
 	config += `	description = "My Interface Description"` + "\n"
 	config += `	shutdown = false` + "\n"
 	config += `	ip_proxy_arp = false` + "\n"
+	config += `	ip_local_proxy_arp = false` + "\n"
 	config += `	ip_redirects = false` + "\n"
 	config += `	ip_unreachables = false` + "\n"
 	config += `	vrf_forwarding = "VRF1"` + "\n"